+++ feature/search/src/main/kotlin/uk/govuk/app/search/ui/widget/SearchWidget.kt
@@ -4,9 +4,9 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import uk.govuk.app.design.ui.component.SearchCard
 import uk.govuk.app.design.ui.theme.GovUkTheme
 import uk.govuk.app.search.R
+import uk.govuk.app.search.ui.component.SearchCard
 
 @Composable
 fun SearchWidget(
